Move PrintGeometryDialog widgets to its own module, make PrintPreviewAction.printPreviewDialog a property
