+++ silx/gui/plot/actions/printpreview.py
@@ -23,17 +23,20 @@
 #
 # ###########################################################################*/
 """
+Print preview action, to send the content of a plot to a print preview page.
+The plot content can then be moved on the page and resized prior to printing.
 
 - :class:`PrintPreviewAction`
 """
 from __future__ import absolute_import
 
-from io import StringIO
 import logging
+from io import StringIO
 
 from . import PlotAction
-from ...widgets.PrintPreview import PrintPreviewDialog
 from ... import qt
+from ...widgets.PrintPreview import PrintPreviewDialog
+from ...widgets.PrintGeometryDialog import PrintGeometryDialog
 
 __authors__ = ["P. Knobel"]
 __license__ = "MIT"
@@ -41,7 +44,6 @@ __date__ = "18/07/2017"
 
 _logger = logging.getLogger(__name__)
 # _logger.setLevel(logging.DEBUG)
-
 
 
 class PrintPreviewAction(PlotAction):
@@ -55,9 +57,8 @@ class PrintPreviewAction(PlotAction):
         super(PrintPreviewAction, self).__init__(
             plot, icon='document-print', text='Print preview',
             tooltip='Send plot data to a print preview dialog',
-            triggered=self._plotToPrintPreview,
-            checkable=False, parent=parent)
-        self.printPreviewDialog = None
+            triggered=self._plotToPrintPreview, parent=parent)
+        self._printPreviewDialog = None
         self.printConfigurationDialog = None
 
         self._printConfiguration = {"xOffset": 0.1,
@@ -67,9 +68,13 @@ class PrintPreviewAction(PlotAction):
                                     "units": "page",
                                     "keepAspectRatio": True}
 
-    def _plotToPrintPreview(self):
+    @property
+    def printPreviewDialog(self):
         if self.printPreviewDialog is None:
-            self.printPreviewDialog = PrintPreviewDialog(self.parent())
+            self._printPreviewDialog = PrintPreviewDialog(self.parent())
+        return self._printPreviewDialog
+
+    def _plotToPrintPreview(self):
         self.printPreviewDialog.ensurePrinterIsSet()
 
         if qt.HAS_SVG:
@@ -221,212 +226,6 @@ class PrintPreviewAction(PlotAction):
         self.printConfigurationDialog.setPrintGeometry(self._printConfiguration)
         if self.printConfigurationDialog.exec_():
             self._printConfiguration = self.printConfigurationDialog.getPrintGeometry()
-
-
-class PrintGeometryWidget(qt.QWidget):
-    """Widget to specify the size and aspect ratio of an item
-    before sending it to the print preview dialog.
-
-    Use methods :meth:`setPrintGeometry` and :meth:`getPrintGeometry`
-    to interact with the widget.
-    """
-    def __init__(self, parent=None):
-        super(PrintGeometryWidget, self).__init__(parent)
-        self.mainLayout = qt.QGridLayout(self)
-        self.mainLayout.setContentsMargins(0, 0, 0, 0)
-        self.mainLayout.setSpacing(2)
-        hbox = qt.QWidget()
-        hboxLayout = qt.QHBoxLayout(hbox)
-        hboxLayout.setContentsMargins(0, 0, 0, 0)
-        hboxLayout.setSpacing(2)
-        label = qt.QLabel(self)
-        label.setText("Units")
-        label.setAlignment(qt.Qt.AlignCenter)
-        self._pageButton = qt.QRadioButton()
-        self._pageButton.setText("Page")
-        self._inchButton = qt.QRadioButton()
-        self._inchButton.setText("Inches")
-        self._cmButton = qt.QRadioButton()
-        self._cmButton.setText("Centimeters")
-        self._buttonGroup = qt.QButtonGroup(self)
-        self._buttonGroup.addButton(self._pageButton)
-        self._buttonGroup.addButton(self._inchButton)
-        self._buttonGroup.addButton(self._cmButton)
-        self._buttonGroup.setExclusive(True)
-
-        # units
-        self.mainLayout.addWidget(label, 0, 0, 1, 4)
-        hboxLayout.addWidget(self._pageButton)
-        hboxLayout.addWidget(self._inchButton)
-        hboxLayout.addWidget(self._cmButton)
-        self.mainLayout.addWidget(hbox, 1, 0, 1, 4)
-        self._pageButton.setChecked(True)
-
-        # xOffset
-        label = qt.QLabel(self)
-        label.setText("X Offset:")
-        self.mainLayout.addWidget(label, 2, 0)
-        self._xOffset = qt.QLineEdit(self)
-        validator = qt.QDoubleValidator(None)
-        self._xOffset.setValidator(validator)
-        self._xOffset.setText("0.0")
-        self.mainLayout.addWidget(self._xOffset, 2, 1)
-
-        # yOffset
-        label = qt.QLabel(self)
-        label.setText("Y Offset:")
-        self.mainLayout.addWidget(label, 2, 2)
-        self._yOffset = qt.QLineEdit(self)
-        validator = qt.QDoubleValidator(None)
-        self._yOffset.setValidator(validator)
-        self._yOffset.setText("0.0")
-        self.mainLayout.addWidget(self._yOffset, 2, 3)
-
-        # width
-        label = qt.QLabel(self)
-        label.setText("Width:")
-        self.mainLayout.addWidget(label, 3, 0)
-        self._width = qt.QLineEdit(self)
-        validator = qt.QDoubleValidator(None)
-        self._width.setValidator(validator)
-        self._width.setText("0.5")
-        self.mainLayout.addWidget(self._width, 3, 1)
-
-        # height
-        label = qt.QLabel(self)
-        label.setText("Height:")
-        self.mainLayout.addWidget(label, 3, 2)
-        self._height = qt.QLineEdit(self)
-        validator = qt.QDoubleValidator(None)
-        self._height.setValidator(validator)
-        self._height.setText("0.5")
-        self.mainLayout.addWidget(self._height, 3, 3)
-
-        # aspect ratio
-        self._aspect = qt.QCheckBox(self)
-        self._aspect.setText("Keep screen aspect ratio")
-        self._aspect.setChecked(True)
-        self.mainLayout.addWidget(self._aspect, 4, 1, 1, 2)
-
-    def getPrintGeometry(self):
-        """Return the print geometry dictionary.
-
-        See :meth:`setPrintGeometry` for documentation about the
-        print geometry dictionary."""
-        ddict = {}
-        if self._inchButton.isChecked():
-            ddict['units'] = "inches"
-        elif self._cmButton.isChecked():
-            ddict['units'] = "centimeters"
-        else:
-            ddict['units'] = "page"
-
-        ddict['xOffset'] = float(self._xOffset.text())
-        ddict['yOffset'] = float(self._yOffset.text())
-        ddict['width'] = float(self._width.text())
-        ddict['height'] = float(self._height.text())
-
-        if self._aspect.isChecked():
-            ddict['keepAspectRatio'] = True
-        else:
-            ddict['keepAspectRatio'] = False
-        return ddict
-
-    def setPrintGeometry(self, geometry=None):
-        """Set the print geometry.
-
-        The geometry parameters must be provided as a dictionary with
-        the following keys:
-
-         - *"xOffset"* (float)
-         - *"yOffset"* (float)
-         - *"width"* (float)
-         - *"height"* (float)
-         - *"units"*: possible values *"page", "inch", "cm"*
-         - *"keepAspectRatio"*: *True* or *False*
-
-        If *units* is *"page"*, the values should be floats in [0, 1.]
-        and are interpreted as a fraction of the page width or height.
-
-        :param dict geometry: Geometry parameters, as a dictionary."""
-        if geometry is None:
-            geometry = {}
-        oldDict = self.getPrintGeometry()
-        for key in ["units", "xOffset", "yOffset",
-                    "width", "height", "keepAspectRatio"]:
-            geometry[key] = geometry.get(key, oldDict[key])
-
-        if geometry['units'].lower().startswith("inc"):
-            self._inchButton.setChecked(True)
-        elif geometry['units'].lower().startswith("c"):
-            self._cmButton.setChecked(True)
-        else:
-            self._pageButton.setChecked(True)
-
-        self._xOffset.setText("%s" % float(geometry['xOffset']))
-        self._yOffset.setText("%s" % float(geometry['yOffset']))
-        self._width.setText("%s" % float(geometry['width']))
-        self._height.setText("%s" % float(geometry['height']))
-        if geometry['keepAspectRatio']:
-            self._aspect.setChecked(True)
-        else:
-            self._aspect.setChecked(False)
-
-
-class PrintGeometryDialog(qt.QDialog):
-    """Dialog embedding a :class:`PrintGeometryWidget`.
-
-    Use methods :meth:`setPrintGeometry` and :meth:`getPrintGeometry`
-    to interact with the widget.
-
-    Execute method :meth:`exec_` to run the dialog.
-    The return value of that method is *True* if the geometry was set
-    (*Ok* button clicked) or *False* if the user clicked the *Cancel*
-    button.
-    """
-
-    def __init__(self, parent=None):
-        qt.QDialog.__init__(self, parent)
-        self.setWindowTitle("Set print size preferences")
-        layout = qt.QVBoxLayout(self)
-        layout.setContentsMargins(0, 0, 0, 0)
-        layout.setSpacing(0)
-        self.configurationWidget = PrintGeometryWidget(self)
-        hbox = qt.QWidget(self)
-        hboxLayout = qt.QHBoxLayout(hbox)
-        self.okButton = qt.QPushButton(hbox)
-        self.okButton.setText("Accept")
-        self.okButton.setAutoDefault(False)
-        self.rejectButton = qt.QPushButton(hbox)
-        self.rejectButton.setText("Dismiss")
-        self.rejectButton.setAutoDefault(False)
-        self.okButton.clicked.connect(self.accept)
-        self.rejectButton.clicked.connect(self.reject)
-        hboxLayout.setContentsMargins(0, 0, 0, 0)
-        hboxLayout.setSpacing(2)
-        # hboxLayout.addWidget(qt.HorizontalSpacer(hbox))
-        hboxLayout.addWidget(self.okButton)
-        hboxLayout.addWidget(self.rejectButton)
-        # hboxLayout.addWidget(qt.HorizontalSpacer(hbox))
-        layout.addWidget(self.configurationWidget)
-        layout.addWidget(hbox)
-
-    def setPrintGeometry(self, geometry):
-        """Return the print geometry dictionary.
-
-        See :meth:`PrintGeometryWidget.setPrintGeometry` for documentation on
-        print geometry dictionary.
-
-        :param dict geometry: Print geometry parameters dictionary.
-        """
-        self.configurationWidget.setPrintGeometry(geometry)
-
-    def getPrintGeometry(self):
-        """Return the print geometry dictionary.
-
-        See :meth:`PrintGeometryWidget.setPrintGeometry` for documentation on
-        print geometry dictionary."""
-        return self.configurationWidget.getPrintGeometry()
 
 
 if __name__ == '__main__':
